Feature register

+++ src/main/java/com/astraval/coreflow/modules/Auth/dto/LoginResponse.java
@@ -12,11 +12,11 @@ public class LoginResponse {
     private Integer userId;
     private String roleCode;
     private String landingUrl;
-    private Integer companyId;
+    private Long companyId;
     private String companyName;
     private List<Long> companyIds;
     
-    public LoginResponse(String token, String refreshToken, Integer userId, String roleCode, String landingUrl, Integer companyId, String companyName, List<Long> companyIds) {
+    public LoginResponse(String token, String refreshToken, Integer userId, String roleCode, String landingUrl, Long companyId, String companyName, List<Long> companyIds) {
         this.token = token;
         this.refreshToken = refreshToken;
         this.userId = userId;
@@ -27,7 +27,7 @@ public class LoginResponse {
         this.companyIds = companyIds;
     }
     
-    public LoginResponse(String token, String refreshToken, Integer userId, String roleCode, String landingUrl, Integer companyId, String companyName) {
+    public LoginResponse(String token, String refreshToken, Integer userId, String roleCode, String landingUrl, Long companyId, String companyName) {
         this(token, refreshToken, userId, roleCode, landingUrl, companyId, companyName, null);
     }
 }
